start to move auth flow from Login to Index
